feat: changed double with float

+++ student_grades_average.cpp
@@ -5,14 +5,14 @@ using namespace std;
 
 int main() {
   const int GRADES_TO_READ = 7;
-  double studentGradesTotal = 0;
+  float studentGradesTotal = 0;
   string studentName;
 
   cout << "Ingresa el nombre del alumno: ";
   getline(cin, studentName);
 
   for (int i = 0; i < GRADES_TO_READ; i++) {
-    double inputGrade;
+    float inputGrade;
 
     while (true) {
       cout << "Ingresa el valor de la calificación " << i + 1 << ": ";
@@ -30,7 +30,7 @@ int main() {
     }
   }
 
-  double gradeAverage = studentGradesTotal / GRADES_TO_READ;
+  float gradeAverage = studentGradesTotal / GRADES_TO_READ;
 
   cout << "El promedio del estudiante " << studentName << " es de "
        << gradeAverage << "\n";
